Add more tests for cz check
