+++ commitizen/cz/jira/jira.py
@@ -70,7 +70,7 @@ class JiraSmartCz(BaseCommitizen):
         return "<ignored text> <ISSUE_KEY> <ignored text> #<COMMAND> <optional COMMAND_ARGUMENTS>"  # noqa
 
     def schema_pattern(self) -> str:
-        return r".*[A-Z]{2,}\-[0-9]+( #| .+ #).+( #.+)*"
+        return r".*[A-Z]{2,}\-[0-9]+( #| .* #).+( #.+)*"
 
     def info(self):
         dir_path = os.path.dirname(os.path.realpath(__file__))
