+++ tests/test_cli.py
@@ -28,40 +28,3 @@ def test_version(mocker):
     cli.main()
 
     error_mock.assert_called_once()
-
-
-def test_check_jira_fails(mocker, capsys):
-    testargs = ["cz", "-n", "cz_jira", "check", "--commit-msg-file", "some_file"]
-    mocker.patch.object(sys, "argv", testargs)
-    mocker.patch(
-        "commitizen.commands.check.open",
-        mocker.mock_open(read_data="random message for J-2 #fake_command blah"),
-    )
-    with pytest.raises(SystemExit):
-        cli.main()
-    _, err = capsys.readouterr()
-    assert "commit validation: failed!" in err
-
-
-def test_check_jira_succeeds(mocker, capsys):
-    testargs = ["cz", "-n", "cz_jira", "check", "--commit-msg-file", "some_file"]
-    mocker.patch.object(sys, "argv", testargs)
-    mocker.patch(
-        "commitizen.commands.check.open",
-        mocker.mock_open(read_data="JR-23 #command some arguments etc"),
-    )
-    cli.main()
-    out, _ = capsys.readouterr()
-    assert "Commit validation: successful!" in out
-
-
-def test_check_conventional_commit_succeeds(mocker, capsys):
-    testargs = ["cz", "check", "--commit-msg-file", "some_file"]
-    mocker.patch.object(sys, "argv", testargs)
-    mocker.patch(
-        "commitizen.commands.check.open",
-        mocker.mock_open(read_data="fix(scope): some commit message"),
-    )
-    cli.main()
-    out, _ = capsys.readouterr()
-    assert "Commit validation: successful!" in out
